Rename WholeSceneCoverage to SpatialCoverage & new interface
